positive int

+++ corehq/form_processor/migrations/0069_auto_20180205_0849.py
@@ -15,7 +15,7 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='xforminstancesql',
             name='build_version',
-            field=models.IntegerField(blank=True, null=True),
+            field=models.PositiveSmallIntegerField(blank=True, null=True),
         ),
         migrations.AddField(
             model_name='xforminstancesql',
